removed all seg faults

+++ proseminar/03/heat_stencil.h
@@ -1,7 +1,7 @@
 #ifndef HEAT_STENCIL_H_
 #define HEAT_STENCIL_H_
 
-typedef double value_t;
+typedef float value_t;
 
 #define IDX_3D(x,y,z,nx,ny) ((x)+(y)*(nx)+(z)*(nx)*(ny))
 #define IDX_2D(x,y,nx) IDX_3D(x,y,0,nx,0)
